PROBL-41 Tag is not a pointer anymore
# Conflicts:
#	sources/Deltares.Probabilistic.Wrapper.Test/ProjectBuilder.cs
#	sources/Deltares.Probabilistic.Wrapper/Model/Project.h

+++ sources/Deltares.Probabilistic/Model/Evaluation.h
@@ -15,8 +15,7 @@ namespace Deltares
 			double Z = nan("");
 			double Weight = 1;
 			int Iteration = -1;
-
-			intptr_t Tag = 0;
+			int Tag = 0;
 		};
 	}
 }
